p2p: format
